Add that will build database table when run
