+++ .eslintrc.js
@@ -3,5 +3,6 @@ module.exports = {
     "rules" : {
         "consistent-return" : 0,
         "class-methods-use-this" : 0,
+        "no-console" : 0,
     }
 };
